requirements.txt

#Web3 related imports
#web3 

#modbus
pymodbus

#MQTT 
paho-mqtt

#Other 
datetime

toml